* seems to work better now
There is still lots of extra verbosity when toggling a selection, but is definately much more usable.  I think it has something to do with the refreshing of the div which holds the currently selected items; since it has a role and an aria-label, the screen reader is re-reading that stuff each time it is refreshed. However, can't confirm this yet.

- added role of group and aria-label = "current selection" to the .multiselect-field element (when selected items show up)
- enter and spacebar toggle selection
- selected items represented by buttons (no separate text and remove button )
- input field added for typeahead (not yet implemented)
- pressing downArrow when closed opens and focuses in the list
- when closed, focus returns to input field
- spacebar clicks to remove (can also toggle from within the list)